requirements.txt
livekit
livekit-api>=0.7.0
python-dotenv
requests
numpy

